# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: TabMail API
  version: "2.0"
  description: |
    TabMail is a multi-tenant, API-first disposable email service with both
    inbound (SMTP receive) and outbound (SMTP send) capabilities.

    ## Quick Start for Bot/Automation Users

    ### Step 1: Get an API Key
    ```bash
    # Login to get a JWT token
    curl -X POST https://your-tabmail/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email":"you@example.com","password":"your-password"}'

    # Create an API Key with the JWT
    curl -X POST https://your-tabmail/api/v1/keys \
      -H "Authorization: Bearer <jwt_token>" \
      -H "Content-Type: application/json" \
      -d '{"label":"my-bot","scopes":["mailboxes:write","messages:read","send:write"]}'
    ```

    ### Step 2: Create a Disposable Mailbox
    ```bash
    # Get a suggested address
    curl https://your-tabmail/api/v1/domains/{domain_id}/suggest-address \
      -H "X-API-Key: tabmail_..."

    # Create the mailbox
    curl -X POST https://your-tabmail/api/v1/mailboxes \
      -H "X-API-Key: tabmail_..." \
      -H "Content-Type: application/json" \
      -d '{"full_address":"suggested-addr@domain.com","zone_id":"..."}'
    ```

    ### Step 3: Receive Emails (Poll or SSE)
    ```bash
    # Poll for messages
    curl https://your-tabmail/api/v1/mailbox/addr@domain.com \
      -H "X-API-Key: tabmail_..."

    # Or use Server-Sent Events for real-time
...
      -H "X-API-Key: tabmail_..."
    ```

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - APIKey: []
components:
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key
    MailboxToken:
      type: http
      scheme: bearer
      bearerFormat: TabMail-Mailbox-Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (alias for APIKey)
...
paths:
  /api/v1/token:
    post:
      tags: [Mailboxes]
      summary: Issue a mailbox access token
      operationId: issueMailboxToken
      security: [ {} ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, password]
              properties:
                address: { type: string, example: "user@mail.example.com" }
                password: { type: string }
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/MailboxTokenResponse" }
        "403": { description: Invalid credentials or mailbox access mode }

  # ── Open Domain Resources ───────────────────────────
  /api/v1/resources/domains:
    get:
      tags: [Domains]
      summary: List verified domain resources open to the current viewer
      operationId: listOpenDomains
      security: [ {}, { BearerAuth: [] } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DomainZone" }
  /api/v1/resources/domains/{id}/suggest-address:
    get:
      tags: [Domains]
      summary: Generate a randomized address for an open domain resource
      operationId: suggestOpenAddress
      security: [ {}, { BearerAuth: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: subdomain
          in: query
          required: false
          schema: { type: boolean }
          description: Requires allow_random_subdomains on the open resource
      responses:
        "200":
          description: Suggested randomized mailbox address
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/SuggestedAddress" }
        "403": { description: Resource is not open to this viewer }
        "404": { description: Not found }

  # ── Domains / Zones ─────────────────────────────────
  /api/v1/domains:
    get:
      tags: [Domains]
      summary: List domain zones
      operationId: listDomains
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DomainZone" }
    post:
      tags: [Domains]
      summary: Bind a new domain zone
      operationId: createDomain
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, example: "mail.example.com" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/DomainZone" }
        "409": { description: Domain already exists }
  /api/v1/domains/{id}:
    delete:
      tags: [Domains]
      summary: Delete a domain zone
      operationId: deleteDomain
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
  /api/v1/domains/{id}/verify:
    post:
      tags: [Domains]
      summary: Trigger DNS verification
      operationId: triggerVerify
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Verification info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      domain: { type: string }
                      txt_record: { type: string }
                      is_verified: { type: boolean }
                      mx_verified: { type: boolean }
                      checks: { type: object }
                      hint: { type: string }
  /api/v1/domains/{id}/verification-status:
    get:
      tags: [Domains]
      summary: Get DNS verification status
      operationId: verificationStatus
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Verification details (TXT, MX, SPF, DKIM, DMARC)
  /api/v1/domains/{id}/suggest-address:
    get:
      tags: [Domains]
      summary: Generate a randomized mailbox address suggestion for a domain zone
      operationId: suggestAddress
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: subdomain
          in: query
          required: false
          schema: { type: boolean }
          description: When true, also randomize a child subdomain under the parent zone
      responses:
        "200":
          description: Suggested randomized mailbox address
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/SuggestedAddress" }
        "400": { description: Bad request }
        "403": { description: Forbidden }
        "404": { description: Not found }

  # ── Domain Routes ───────────────────────────────────
  /api/v1/domains/{id}/routes:
    get:
      tags: [Domain Routes]
      summary: List routes for a zone
      operationId: listRoutes
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DomainRoute" }
    post:
      tags: [Domain Routes]
      summary: Create a routing rule
      operationId: createRoute
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [route_type, match_value]
              properties:
                route_type: { type: string, enum: [exact, wildcard, deep_wildcard, sequence] }
                match_value: { type: string, example: "*.mail.example.com" }
                range_start: { type: integer }
                range_end: { type: integer }
                auto_create_mailbox: { type: boolean, default: true }
                retention_hours_override: { type: integer }
                access_mode_default: { type: string, enum: [public, token, api_key], default: public }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/DomainRoute" }
  /api/v1/domains/{id}/routes/{routeId}:
    delete:
      tags: [Domain Routes]
      summary: Delete a route
      operationId: deleteRoute
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: routeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  # ── Mailboxes ───────────────────────────────────────
  /api/v1/mailboxes:
    get:
      tags: [Mailboxes]
      summary: List mailboxes
      operationId: listMailboxes
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Mailbox" }
                  meta: { $ref: "#/components/schemas/Meta" }
    post:
      tags: [Mailboxes]
      summary: Create a mailbox
      operationId: createMailbox
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address: { type: string, example: "user@mail.example.com" }
                password: { type: string }
                access_mode: { type: string, enum: [public, token, api_key], default: public }
                retention_hours_override: { type: integer, nullable: true, example: 24 }
                expires_at: { type: string, format: date-time, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Mailbox" }
        "409": { description: Address already exists }
  /api/v1/mailboxes/{id}:
    delete:
      tags: [Mailboxes]
      summary: Delete a mailbox
      operationId: deleteMailbox
      security: [{ APIKey: [] }, { BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  # ── Messages ────────────────────────────────────────
  /api/v1/mailbox/{address}:
    get:
      tags: [Messages]
      summary: List messages in a mailbox
      operationId: listMessages
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
          example: "user@mail.example.com"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
                  meta: { $ref: "#/components/schemas/Meta" }
    delete:
      tags: [Messages]
      summary: Purge all messages from a mailbox
      operationId: purgeMailbox
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Purged }
  /api/v1/mailbox/{address}/events:
    get:
      tags: [Messages]
      summary: Stream mailbox events with SSE
      operationId: streamMailboxEvents
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/mailbox/{address}/{id}:
    get:
      tags: [Messages]
      summary: Get message detail
      operationId: getMessage
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/MessageDetail" }
    patch:
      tags: [Messages]
      summary: Mark message as seen
      operationId: markSeen
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      seen: { type: boolean }
    delete:
      tags: [Messages]
      summary: Delete a single message
      operationId: deleteMessage
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
  /api/v1/mailbox/{address}/{id}/source:
    get:
      tags: [Messages]
      summary: Get raw RFC 822 source
      operationId: getMessageSource
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Raw email
          content:
            message/rfc822:
              schema: { type: string, format: binary }

  # ── Sending ─────────────────────────────────────────
  /api/v1/send:
    post:
      tags: [Sending]
      summary: Send an email
      operationId: sendEmail
      description: |
        Send an outbound email. Requires `send:write` scope for API keys,
        or JWT authentication with `can_send` permission enabled.

        The sender (`from`) domain must be a verified domain zone owned by
        the caller's tenant.

        **Rate limiting**: Subject to daily send quota configured in the
        user's permission profile.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, to, subject]
              properties:
                from:
                  type: string
                  format: email
                  description: Sender address. Domain must be a verified zone.
                  example: "noreply@mail.example.com"
                to:
                  type: array
                  items:
                    type: string
                    format: email
                  minItems: 1
                  description: Primary recipients
                  example: ["user@gmail.com"]
                cc:
                  type: array
                  items:
                    type: string
                    format: email
                  description: CC recipients
                bcc:
                  type: array
                  items:
                    type: string
                    format: email
                  description: BCC recipients (hidden from other recipients)
                subject:
                  type: string
                  maxLength: 998
                  description: Email subject line
                  example: "Your verification code"
                text_body:
                  type: string
                  description: Plain text body
                  example: "Your code is 123456"
                html_body:
                  type: string
                  description: HTML body
                  example: "<p>Your code is <b>123456</b></p>"
                headers:
                  type: object
                  additionalProperties:
                    type: string
                  description: Custom email headers
                  example:
                    Reply-To: "support@example.com"
      responses:
        '201':
          description: Email queued for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendEmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/QuotaExceeded'
  /api/v1/outbound:
    get:
      tags: [Sending]
      summary: List outbound jobs
      operationId: listOutboundJobs
      description: List outbound email delivery jobs for the current tenant.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Paginated list of outbound jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OutboundJob'
                  total:
                    type: integer
  /api/v1/outbound/{id}:
    get:
      tags: [Sending]
      summary: Get outbound job status
      operationId: getOutboundJob
      description: Get the delivery status of a specific outbound email job.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Outbound job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundJob'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Admin: Domain Resources ─────────────────────────
  /api/v1/admin/domains:
    get:
      tags: [Admin]
      summary: List all domain resources
      operationId: adminListDomains
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DomainZone" }
  /api/v1/admin/domains/{id}:
    patch:
      tags: [Admin]
      summary: Update domain resource visibility and random subdomain policy
      operationId: adminUpdateDomainAccess
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility: { type: string, enum: [private, authenticated, public] }
                allow_random_subdomains: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/DomainZone" }
        "400": { description: Invalid visibility or random-subdomain state }
        "403": { description: Admin access required }
        "404": { description: Not found }

  # ── Admin: Permissions ─────────────────────────────
...
  /api/v1/auth/me/permissions:
    get:
      tags: [Permissions]
      summary: Get my permissions
      operationId: getMyPermissions
      description: Get the effective permissions for the currently authenticated user.
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/EffectivePermission" }
        "401": { description: Not authenticated }

  # ── Admin: Tenants ──────────────────────────────────
...
  /api/v1/admin/status:
    get:
      tags: [Admin]
      summary: System status alias
      operationId: getStatus
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
...